Abstract ui logic out of metricCalculator
