+++ src/main/java/seedu/address/ui/MetricsWindow.java
@@ -3,6 +3,8 @@ package seedu.address.ui;
 import java.util.logging.Logger;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+import javafx.scene.control.Separator;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import seedu.address.commons.core.LogsCenter;
@@ -89,7 +91,72 @@ public class MetricsWindow extends ClosableWindow {
         }
 
         MetricsCalculator.MetricsData metricsData = metricsCalculator.calculateMetrics(currentAddressBook);
-        metricsCalculator.renderMetrics(statusMetrics, metricsData);
+        renderMetrics(statusMetrics, metricsData);
+    }
+
+    /**
+     * Renders the metrics data into the provided VBox container.
+     * Clears existing content and populates with formatted metrics display.
+     *
+     * @param container The VBox container to render metrics into
+     * @param metricsData The calculated metrics data to display
+     */
+    private void renderMetrics(VBox container, MetricsCalculator.MetricsData metricsData) {
+        container.getChildren().clear();
+
+        if (!metricsData.hasData()) {
+            renderNoDataMessage(container);
+            return;
+        }
+
+        renderTotalCount(container, metricsData.getTotalCompanies());
+        renderSeparator(container);
+        renderStatusMetrics(container, metricsData);
+    }
+
+    /**
+     * Renders a "no data" message when no companies are found.
+     */
+    private void renderNoDataMessage(VBox container) {
+        Label noDataLabel = new Label("No companies found");
+        noDataLabel.getStyleClass().add("metrics-no-data");
+        container.getChildren().add(noDataLabel);
+    }
+
+    /**
+     * Renders the total company count.
+     */
+    private void renderTotalCount(VBox container, long totalCompanies) {
+        Label totalLabel = new Label(String.format("Total Companies: %d", totalCompanies));
+        totalLabel.getStyleClass().add("metrics-total");
+        container.getChildren().add(totalLabel);
+    }
+
+    /**
+     * Renders a visual separator.
+     */
+    private void renderSeparator(VBox container) {
+        Separator separator = new Separator();
+        separator.getStyleClass().add("metrics-separator");
+        container.getChildren().add(separator);
+    }
+
+    /**
+     * Renders individual status metrics in the defined order.
+     */
+    private void renderStatusMetrics(VBox container, MetricsCalculator.MetricsData metricsData) {
+        for (String status : metricsData.getStatusOrder()) {
+            long count = metricsData.getStatusCount(status);
+            double percentage = metricsData.getStatusPercentage(status);
+
+            String displayText = String.format("%s:  %d (%.1f%%)", status, count, percentage);
+
+            Label statusLabel = new Label(displayText);
+            String statusUpperCase = status.toUpperCase().replace('_', '-');
+            statusLabel.getStyleClass().addAll("metrics-status",
+                    "status-" + statusUpperCase, "status-color-" + statusUpperCase);
+            container.getChildren().add(statusLabel);
+        }
     }
